Plugin for organization course files aggregation
The process written in src/utils/coursesLoader.js is difficult to internationalise due to the use of require.export() and webpack's bundling process. It is therefore easier to generate a main file that gathers the content of all those files.
So the plugin "organization-courses-agregator-plugin" is use in the docusaurus lifecycle APIs (`loadContent()` by now).
This is a very primitive under-optimized working version